requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
django-extensions==3.2.0
django-guardian==2.4.0
djangorestframework==3.13.1
djangorestframework-bulk==0.2.1
djangorestframework-simplejwt==5.2.0
Pygments==2.12.0
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
drf-yasg
